+++ crd-generator/api/src/main/java/io/fabric8/crd/generator/AbstractJsonSchema.java
@@ -101,6 +101,7 @@ public abstract class AbstractJsonSchema<T, B> {
   public static final String ANNOTATION_JSON_IGNORE = "com.fasterxml.jackson.annotation.JsonIgnore";
   public static final String ANNOTATION_JSON_ANY_GETTER = "com.fasterxml.jackson.annotation.JsonAnyGetter";
   public static final String ANNOTATION_JSON_ANY_SETTER = "com.fasterxml.jackson.annotation.JsonAnySetter";
+  public static final String ANNOTATION_DEFAULT = "io.fabric8.generator.annotation.Default";
   public static final String ANNOTATION_MIN = "io.fabric8.generator.annotation.Min";
   public static final String ANNOTATION_MAX = "io.fabric8.generator.annotation.Max";
   public static final String ANNOTATION_PATTERN = "io.fabric8.generator.annotation.Pattern";
@@ -143,6 +144,7 @@ public abstract class AbstractJsonSchema<T, B> {
   }
 
   protected static class SchemaPropsOptions {
+    final String defaultValue;
     final Double min;
     final Double max;
     final String pattern;
@@ -152,6 +154,7 @@ public abstract class AbstractJsonSchema<T, B> {
     final boolean preserveUnknownFields;
 
     SchemaPropsOptions() {
+      defaultValue = null;
       min = null;
       max = null;
       pattern = null;
@@ -160,14 +163,19 @@ public abstract class AbstractJsonSchema<T, B> {
       preserveUnknownFields = false;
     }
 
-    public SchemaPropsOptions(Double min, Double max, String pattern,
+    public SchemaPropsOptions(String defaultValue, Double min, Double max, String pattern,
         boolean nullable, boolean required, boolean preserveUnknownFields) {
+      this.defaultValue = defaultValue;
       this.min = min;
       this.max = max;
       this.pattern = pattern;
       this.nullable = nullable;
       this.required = required;
       this.preserveUnknownFields = preserveUnknownFields;
+    }
+
+    public Optional<String> getDefault() {
+      return Optional.ofNullable(defaultValue);
     }
 
     public Optional<Double> getMin() {
@@ -320,6 +328,7 @@ public abstract class AbstractJsonSchema<T, B> {
       }
 
       SchemaPropsOptions options = new SchemaPropsOptions(
+          facade.defaultValue,
           facade.min,
           facade.max,
           facade.pattern,
@@ -349,6 +358,7 @@ public abstract class AbstractJsonSchema<T, B> {
     private final String propertyName;
     private final String type;
     private String renamedTo;
+    private String defaultValue;
     private Double min;
     private Double max;
     private String pattern;
@@ -377,6 +387,9 @@ public abstract class AbstractJsonSchema<T, B> {
     public void process() {
       annotations.forEach(a -> {
         switch (a.getClassRef().getFullyQualifiedName()) {
+          case ANNOTATION_DEFAULT:
+            defaultValue = (String) a.getParameters().get(VALUE);
+            break;
           case ANNOTATION_NULLABLE:
             nullable = true;
             break;
@@ -425,6 +438,10 @@ public abstract class AbstractJsonSchema<T, B> {
 
     public boolean isNullable() {
       return nullable;
+    }
+
+    public Optional<String> getDefault() {
+      return Optional.ofNullable(defaultValue);
     }
 
     public Optional<Double> getMax() {
@@ -481,6 +498,7 @@ public abstract class AbstractJsonSchema<T, B> {
     private final List<PropertyOrAccessor> propertyOrAccessors = new ArrayList<>(4);
     private String renamedTo;
     private String description;
+    private String defaultValue;
     private Double min;
     private Double max;
     private String pattern;
@@ -511,6 +529,7 @@ public abstract class AbstractJsonSchema<T, B> {
         propertyOrAccessors.add(PropertyOrAccessor.fromMethod(method, name));
       }
       schemaFrom = schemaSwap;
+      defaultValue = null;
       min = null;
       max = null;
       pattern = null;
@@ -539,6 +558,7 @@ public abstract class AbstractJsonSchema<T, B> {
             LOGGER.debug("Description for property {} has already been contributed by: {}", name, descriptionContributedBy);
           }
         }
+        defaultValue = p.getDefault().orElse(null);
         min = p.getMin().orElse(min);
         max = p.getMax().orElse(max);
         pattern = p.getPattern().orElse(pattern);