+++ crd-generator/api/src/main/java/io/fabric8/crd/generator/CRDGenerator.java
@@ -58,7 +58,7 @@ public class CRDGenerator {
   private boolean parallel;
   private Map<String, CustomResourceInfo> infos;
 
-  private static final ObjectMapper YAML_MAPPER = JsonMapper.builder(new YAMLFactory()
+  public static final ObjectMapper YAML_MAPPER = JsonMapper.builder(new YAMLFactory()
       .enable(YAMLGenerator.Feature.MINIMIZE_QUOTES)
       .enable(YAMLGenerator.Feature.ALWAYS_QUOTE_NUMBERS_AS_STRINGS)
       .disable(YAMLGenerator.Feature.WRITE_DOC_START_MARKER))