+++ crd-generator/api/src/test/java/io/fabric8/crd/example/annotated/AnnotatedSpec.java
@@ -18,6 +18,7 @@ package io.fabric8.crd.example.annotated;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyDescription;
+import io.fabric8.generator.annotation.Default;
 import io.fabric8.generator.annotation.Max;
 import io.fabric8.generator.annotation.Min;
 import io.fabric8.generator.annotation.Nullable;
@@ -35,6 +36,7 @@ public class AnnotatedSpec {
   private int max;
   private String singleDigit;
   private String nullable;
+  private String defaultValue;
   @Required
   private boolean emptySetter;
   @Required
@@ -82,6 +84,11 @@ public class AnnotatedSpec {
   @Nullable
   public String getNullable() {
     return null;
+  }
+
+  @Default("my-value")
+  public String getDefaultValue() {
+    return "foo";
   }
 
   @JsonProperty